babyshop_app/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
flake8==7.1.0
gunicorn==22.0.0
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
pycodestyle==2.12.0
pyflakes==3.2.0
sqlparse==0.5.0
